backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
celery==5.3.4
redis==5.0.1
websockets==12.0
python-socketio==5.10.0
aiofiles==23.2.1
httpx==0.25.2
jinja2==3.1.2
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
loguru==0.7.2
prometheus-client==0.19.0